refactor: 시스템 프롬프트 파일 분리, DTO 분리, JSON 추출 개선
- 시스템 프롬프트를 resources/prompts/event-filter-system.txt로 분리
- LLM 요청/응답 DTO를 dto/LlmApiDto.kt로 분리
- JSON 추출을 indexOf/lastIndexOf 방식으로 변경하여 LLM 응답 안정성 향상
